refactor: update path resolution for @ alias
