+++ app/components/dialog/delete-dialog/delete-dialog.test.tsx
@@ -1,7 +1,7 @@
 import { render, RenderResult } from '@testing-library/react'
 import { afterAll, beforeAll, describe, expect, test } from 'vitest'
 
-import { DeleteDialog } from '@/app/components/dialog/delete-dialog/delete-dialog'
+import { DeleteDialog } from '@/components/dialog/delete-dialog/delete-dialog'
 
 test('DeleteDialog', () => {
   describe('DeleteDialog', () => {
