+++ app/components/menu-bar/components/login-button/login-button.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, RenderResult } from '@testing-library/react'
 import { afterEach, beforeAll, beforeEach, describe, expect, test } from 'vitest'
 
-import { LoginButton } from '@/app/components/menu-bar/components/login-button/login-button'
+import { LoginButton } from '@/components/menu-bar/components/login-button/login-button'
 
 test('LoginButton', () => {
   let LoginButtonMounted: RenderResult
