+++ app/components/menu-bar/components/user-dropdown/user-dropdown.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, RenderResult, screen } from '@testing-library/react'
 import { signOut } from 'next-auth/react'
 import { afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest'
 
-import { UserDropdown } from '@/app/components/menu-bar/components/user-dropdown/user-dropdown'
+import { UserDropdown } from '@/components/menu-bar/components/user-dropdown/user-dropdown'
 
 vi.mock('next-auth/react', () => ({
   signOut: vi.fn(),
