+++ app/components/menu-bar/menu-bar.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, RenderResult, screen } from '@testing-library/react'
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 
-import { MenuBar } from '@/app/components/menu-bar/menu-bar'
+import { MenuBar } from '@/components/menu-bar/menu-bar'
 
 vi.mock('next-auth/react', () => ({
   useSession: () => ({ status: 'unauthenticated', data: null }),
