+++ app/components/retro-section/retro-section.tsx
@@ -6,9 +6,9 @@ import { useTheme } from 'next-themes'
 import { useEffect, useState } from 'react'
 import GridLoader from 'react-spinners/GridLoader'
 
-import { RetroDialog } from '@/app/components/retro-section/components/retro-dialog'
-import { RetrospectiveCreateInput } from '@/app/types/retrospective'
-import { trpc } from '@/app/utils/trpc'
+import { RetroDialog } from '@/components/retro-section/components/retro-dialog'
+import { RetrospectiveCreateInput } from '@/types/retrospective'
+import { trpc } from '@/utils/trpc'
 
 type RetroSectionProps = {
   userId: User['id']
