+++ app/components/retro-view/components/action-bar/components/phase-indicator/phase-indicator.test.tsx
@@ -2,7 +2,7 @@ import { Retrospective } from '@prisma/client'
 import { fireEvent, render, screen } from '@testing-library/react'
 import { expect, test, vi } from 'vitest'
 
-import { PhaseIndicator } from '@/app/components/retro-view/components/action-bar/components/phase-indicator/phase-indicator'
+import { PhaseIndicator } from '@/components/retro-view/components/action-bar/components/phase-indicator/phase-indicator'
 
 test('PhaseIndicator', () => {
   const handleUpdateRetroMock = vi.fn()
