+++ app/components/retro-view/components/retro-action-bar.tsx
@@ -3,10 +3,10 @@
 import { Retrospective } from '@prisma/client'
 import { useState } from 'react'
 
-import { ActionButtons } from '@/app/components/retro-view/components/action-bar/components/action-buttons/action-buttons'
-import { PhaseIndicator } from '@/app/components/retro-view/components/action-bar/components/phase-indicator/phase-indicator'
-import { RetroTimer } from '@/app/components/retro-view/components/action-bar/components/retro-timer/retro-timer'
-import { trpc } from '@/app/utils/trpc'
+import { ActionButtons } from '@/components/retro-view/components/action-bar/components/action-buttons/action-buttons'
+import { PhaseIndicator } from '@/components/retro-view/components/action-bar/components/phase-indicator/phase-indicator'
+import { RetroTimer } from '@/components/retro-view/components/action-bar/components/retro-timer/retro-timer'
+import { trpc } from '@/utils/trpc'
 
 type RetroActionBarProps = {
   selectedRetro: Retrospective
