+++ app/components/retro-view/views/discussing-phase/components/item-stack/item-stack.test.tsx
@@ -2,7 +2,7 @@ import { RetroItem } from '@prisma/client'
 import { render, screen } from '@testing-library/react'
 import { beforeEach, describe, expect, test } from 'vitest'
 
-import { ItemStack } from '@/app/components/retro-view/views/discussing-phase/components/item-stack/item-stack'
+import { ItemStack } from '@/components/retro-view/views/discussing-phase/components/item-stack/item-stack'
 
 test('ItemStack', () => {
   describe('ItemStack', () => {
