+++ app/components/retro-view/views/voting-phase/components/item-voter.test.tsx
@@ -2,14 +2,14 @@ import { RetroItem, Retrospective } from '@prisma/client'
 import { fireEvent, render, screen } from '@testing-library/react'
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 
-import { ItemVoter } from '@/app/components/retro-view/views/voting-phase/components/item-voter'
-import { trpc } from '@/app/utils/trpc'
+import { ItemVoter } from '@/components/retro-view/views/voting-phase/components/item-voter'
+import { trpc } from '@/utils/trpc'
 
 vi.mock('next-auth/react', () => ({
   useSession: vi.fn(),
 }))
 
-vi.mock('@/app/utils/trpc', () => ({
+vi.mock('@/utils/trpc', () => ({
   trpc: {
     retroItem: {
       getAllByRetroId: {
