+++ app/components/retro-view/views/writing-phase/components/item-collector/item-collector.tsx
@@ -3,11 +3,11 @@
 import { RetroItem, Retrospective } from '@prisma/client'
 import { useSession } from 'next-auth/react'
 
-import { DeleteDialog } from '@/app/components/dialog/delete-dialog/delete-dialog'
-import { EditDialog } from '@/app/components/dialog/edit-dialog/edit-dialog'
-import { RetroItemDialog } from '@/app/components/retro-view/views/writing-phase/components/item-collector/components/retro-item-dialog'
-import { RetroItemCreateInput } from '@/app/types/retro-item'
-import { trpc } from '@/app/utils/trpc'
+import { DeleteDialog } from '@/components/dialog/delete-dialog/delete-dialog'
+import { EditDialog } from '@/components/dialog/edit-dialog/edit-dialog'
+import { RetroItemDialog } from '@/components/retro-view/views/writing-phase/components/item-collector/components/retro-item-dialog'
+import { RetroItemCreateInput } from '@/types/retro-item'
+import { trpc } from '@/utils/trpc'
 
 type ItemCollectorProps = {
   title: string
