+++ app/components/sign-up/sign-up-form.test.tsx
@@ -1,7 +1,7 @@
 import { render, RenderResult } from '@testing-library/react'
 import { afterAll, beforeAll, describe, test } from 'vitest'
 
-import { SignUpForm } from '@/app/components/sign-up/sign-up-form'
+import { SignUpForm } from '@/components/sign-up/sign-up-form'
 
 test('SignUpForm', () => {
   describe('SignUpForm', () => {
