+++ app/components/team-section/components/invite-user-dialog.test.tsx
@@ -1,7 +1,7 @@
 import { render, RenderResult } from '@testing-library/react'
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 
-import { InviteUserDialog } from '@/app/components/team-section/components/invite-user-dialog'
+import { InviteUserDialog } from '@/components/team-section/components/invite-user-dialog'
 
 // Mock next-auth/react signIn function
 vi.mock('next-auth/react', () => ({
