+++ app/components/team-section/team-section.tsx
@@ -7,8 +7,8 @@ import { useTheme } from 'next-themes'
 import { useState } from 'react'
 import { GridLoader } from 'react-spinners'
 
-import { InviteUserDialog } from '@/app/components/team-section/components/invite-user-dialog'
-import { trpc } from '@/app/utils/trpc'
+import { InviteUserDialog } from '@/components/team-section/components/invite-user-dialog'
+import { trpc } from '@/utils/trpc'
 
 export function TeamSection() {
   const { resolvedTheme } = useTheme()
