+++ app/server/auth.ts
@@ -1,7 +1,7 @@
 import type { GetServerSidePropsContext } from 'next'
 import { getServerSession } from 'next-auth'
 
-import { authOptions as nextAuthOptions } from '@/pages/api/auth/[...nextauth]'
+import { authOptions as nextAuthOptions } from './../../pages/api/auth/[...nextauth]'
 
 // Next API route example - /pages/api/restricted.ts
 export const getServerAuthSession = async (ctx: {
