+++ app/server/db.ts
@@ -1,7 +1,7 @@
 // src/server/db/client.ts
 import { PrismaClient } from '@prisma/client'
 
-import { env } from '@/app/env/server.mjs'
+import { env } from '@/env/server.mjs'
 
 declare global {
   // eslint-disable-next-line no-var
