+++ pages/dashboard.tsx
@@ -3,11 +3,11 @@ import type { NextPage } from 'next'
 import Head from 'next/head'
 import { useSession } from 'next-auth/react'
 
-import { Feedback } from '@/app/components/feedback/feedback'
-import { MenuBar } from '@/app/components/menu-bar/menu-bar'
-import { RetroSection } from '@/app/components/retro-section/retro-section'
-import { SignUpForm } from '@/app/components/sign-up/sign-up-form'
-import { TeamSection } from '@/app/components/team-section/team-section'
+import { Feedback } from '@/components/feedback/feedback'
+import { MenuBar } from '@/components/menu-bar/menu-bar'
+import { RetroSection } from '@/components/retro-section/retro-section'
+import { SignUpForm } from '@/components/sign-up/sign-up-form'
+import { TeamSection } from '@/components/team-section/team-section'
 
 const Dashboard: NextPage = () => {
   const { data: session, status } = useSession()
