+++ pages/index.tsx
@@ -14,8 +14,8 @@ import Link from 'next/link'
 import { signIn, useSession } from 'next-auth/react'
 import { useTheme } from 'next-themes'
 
-import { Footer } from '@/app/components/footer/footer'
-import { MenuBar } from '@/app/components/menu-bar/menu-bar'
+import { Footer } from '@/components/footer/footer'
+import { MenuBar } from '@/components/menu-bar/menu-bar'
 
 const Landingpage: NextPage = () => {
   const theme = useTheme()
