+++ pages/retro.tsx
@@ -3,10 +3,10 @@ import { NextPage } from 'next'
 import Head from 'next/head'
 import { useSession } from 'next-auth/react'
 
-import { Feedback } from '@/app/components/feedback/feedback'
-import { MenuBar } from '@/app/components/menu-bar/menu-bar'
-import { RetroView } from '@/app/components/retro-view/retro-view'
-import { SignUpForm } from '@/app/components/sign-up/sign-up-form'
+import { Feedback } from '@/components/feedback/feedback'
+import { MenuBar } from '@/components/menu-bar/menu-bar'
+import { RetroView } from '@/components/retro-view/retro-view'
+import { SignUpForm } from '@/components/sign-up/sign-up-form'
 
 const Retro: NextPage = () => {
   const { data: session, status } = useSession()
