+++ pages/terms.tsx
@@ -1,8 +1,8 @@
 import type { NextPage } from 'next'
 import Head from 'next/head'
 
-import { Footer } from '@/app/components/footer/footer'
-import { MenuBar } from '@/app/components/menu-bar/menu-bar'
+import { Footer } from '@/components/footer/footer'
+import { MenuBar } from '@/components/menu-bar/menu-bar'
 
 const Terms: NextPage = () => {
   return (
